Move class into utils
It’s shared between multiple views so shouldn’t sit in one or the other
