+++ tests/file_checks/test_views.py
@@ -16,7 +16,7 @@ def _file_checks(client, file_content):
 @pytest.fixture
 def antivirus(mocker):
     return mocker.patch(
-        "app.file_checks.views.antivirus_client",
+        "app.utils.file_checks.antivirus_client",
         # prevent LocalProxy being detected as an async function
         new_callable=mocker.MagicMock,
     )
